Refactor: Rename __smart_json__.py to core.py and update imports
I renamed the main module file `__smart_json__.py` to `core.py` for clarity and better representation of its role.
I updated all internal project imports in `__init__.py`, test files, and example scripts to reflect this change.

I verified that all unit tests pass and example scripts run as expected after the renaming and import updates (barring pre-existing, unrelated issues in one example).
