+++ output_examples/article_no_schema.json
@@ -1,0 +1,8 @@
+{
+  "Article": {
+    "author": "Dr. AI Expert",
+    "content_summary": "An insightful look into upcoming AI trends.",
+    "publication_date": "2024-03-10",
+    "title": "The Future of AI"
+  }
+}
