+++ output_examples/malformed_article.json
@@ -1,0 +1,5 @@
+
+{
+    "title": "Incomplete Article",
+    "publication_date": "2024-03-11"
+}
