+++ smartjson/__init__.py
@@ -1,2 +1,2 @@
 from __future__ import print_function, unicode_literals, division, absolute_import
-from .__smart_json__ import SmartJson
+from .core import SmartJson
