+++ smartjson/core.py
